chore: download binary file on first run
